style: LARGE refactor
