--- src/App.module.scss
@@ -1,119 +1,0 @@
-.container {
-  height: 100vh;
-  width: 100vw;
-}
-
-.search {
-  padding: var(--size-xLarge);
-  font-size: var(--size-xxLarge);
-  width: 100%;
-  box-sizing: border-box;
-  outline: none;
-  border: none;
-  background: var(--background);
-  color: var(--foreground);
-}
-
-.animator {
-  max-width: 600px;
-  width: 100%;
-  background: var(--background);
-  color: var(--foreground);
-  border-radius: var(--size-large);
-  overflow: hidden;
-  box-shadow: var(--shadow);
-}
-
-.groupName {
-  padding: var(--size-large) var(--size-xxLarge);
-  font-size: 10px;
-  text-transform: uppercase;
-  opacity: 0.5;
-  background: var(--background);
-}
-
-.text {
-  position: sticky;
-  height: 80vh;
-  align-items: center;
-  display: flex;
-  justify-content: center;
-}
-
-.resultItem {
-  padding: var(--size-xLarge) var(--size-xxLarge);
-  background: var(--background);
-  border-left: var(--size-xSmall) solid transparent;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  cursor: pointer;
-
-  &Active {
-    background: var(--a1) !important;
-    border-left: var(--size-xSmall) solid var(--foreground) !important;
-  }
-
-  div {
-    display: flex;
-  }
-
-  &Icon {
-    gap: var(--size-large);
-    align-items: center;
-    &Subtitle {
-      flex-direction: column;
-    }
-  }
-
-  &ShortCut {
-    display: grid;
-    grid-auto-flow: column;
-    gap: var(--size-small);
-
-    kbd {
-      padding: var(--size-small) var(--size-medium);
-      background: rgba(#000, 0.1);
-      border-radius: var(--size-small);
-    }
-  }
-}
-
-.name {
-  position: absolute;
-  display: flex;
-  height: 19.5vh;
-  align-items: flex-end;
-  margin-left: 0.5vw;
-}
-
-.empty {
-  opacity: 0;
-}
-
-.resume {
-  margin-top: var(--size-medium) !important;
-  margin-bottom: var(--size-medium) !important;
-  display: flex;
-  justify-content: center;
-  :global(.annotationLayer) {
-    max-width: 0 !important;
-    max-height: 0 !important;
-  }
-}
-
-.modal {
-  position: absolute;
-  left: 0;
-  top: 5vh;
-  width: 100vw !important;
-
-  canvas {
-    box-shadow: 0 0 var(--size-medium) var(--foreground) !important;
-    border-radius: var(--size-medium) !important;
-  }
-}
-
-.modalContent {
-  overflow: hidden;
-}
