--- src/App.module.scss.d.ts
@@ -1,15 +1,0 @@
-export const animator: string;
-export const container: string;
-export const empty: string;
-export const groupName: string;
-export const modal: string;
-export const modalContent: string;
-export const name: string;
-export const resultItem: string;
-export const resultItemActive: string;
-export const resultItemIcon: string;
-export const resultItemIconSubtitle: string;
-export const resultItemShortCut: string;
-export const resume: string;
-export const search: string;
-export const text: string;
