+++ src/components/KBar/KBar.module.scss.d.ts
@@ -1,0 +1,2 @@
+export const animator: string;
+export const search: string;
