+++ src/components/KBar/Results/Results.module.scss.d.ts
@@ -1,0 +1,6 @@
+export const groupName: string;
+export const resultItem: string;
+export const resultItemActive: string;
+export const resultItemIcon: string;
+export const resultItemIconSubtitle: string;
+export const resultItemShortCut: string;
