+++ src/components/Resume/Resume.module.scss.d.ts
@@ -1,0 +1,4 @@
+export const document: string;
+export const mask: string;
+export const modal: string;
+export const modalContent: string;
