+++ src/index.tsx
@@ -1,7 +1,6 @@
 import React from "react";
 import { createRoot } from "react-dom/client";
-import App from "./App";
-import reportWebVitals from "./reportWebVitals";
+import Home from "@Pages/home";
 import "./index.scss";
 import { BrowserRouter as Router } from "react-router-dom";
 import { H } from "highlight.run";
@@ -15,12 +14,7 @@ const root = createRoot(container!);
 root.render(
   <React.StrictMode>
     <Router>
-      <App />
+      <Home />
     </Router>
   </React.StrictMode>
 );
-
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-reportWebVitals();
