+++ src/pages/home.module.scss
@@ -1,0 +1,15 @@
+.text {
+  position: sticky;
+  height: 80vh;
+  align-items: center;
+  display: flex;
+  justify-content: center;
+}
+
+.name {
+  position: absolute;
+  display: flex;
+  height: 19.5vh;
+  align-items: flex-end;
+  margin-left: 0.5vw;
+}
