+++ src/pages/home.module.scss.d.ts
@@ -1,0 +1,2 @@
+export const name: string;
+export const text: string;
