+++ src/utils/Utils.ts
@@ -1,0 +1,6 @@
+export const getOs = () => {
+  if (navigator.userAgent.includes("Mac")) {
+    return "⌘+K";
+  }
+  return "ctrl+K";
+};
